Elevator Work and Files Cleanup
The elevator works when it loads and then when used again on another floor, it breaks and then when you use it again it works but the animation of the floor you choose the second time stays open. (gotta fix this soon)

I also cleaned up your files in the Project, the other folders I left outside of `#Extra` are things I don't know where to put but I'm sure you'll figure it out. Hope I didn't ruin your file management. lol